Enhance platform management in Group and LocalRuntimeFactory; add platform retrieval and application file handling

+++ agent/src/main/java/dev/httpmarco/polocloud/agent/runtime/local/LocalRuntimeFactory.kt
@@ -4,11 +4,14 @@ import dev.httpmarco.polocloud.agent.Agent
 import dev.httpmarco.polocloud.agent.logger
 import dev.httpmarco.polocloud.agent.runtime.RuntimeFactory
 import dev.httpmarco.polocloud.agent.services.Service
+import java.nio.file.Files
+import java.nio.file.StandardCopyOption
 import kotlin.io.path.ExperimentalPathApi
 import kotlin.io.path.Path
 import kotlin.io.path.createDirectories
 import kotlin.io.path.deleteRecursively
 import kotlin.io.path.exists
+import kotlin.io.path.name
 
 class LocalRuntimeFactory : RuntimeFactory {
 
@@ -25,18 +28,31 @@ class LocalRuntimeFactory : RuntimeFactory {
 
     override fun bootApplication(service: Service) {
 
-        if(service.state != Service.State.PREPARING) {
+        if (service.state != Service.State.PREPARING) {
             logger.error("Cannot boot application for service ${service.name()} because it is not in PREPARING state, but in ${service.state} state&8. &7Wait for action&8...")
             return
         }
 
         logger.info("The service &3${service.name()}&7 is now starting&8...")
 
+        val platform = service.group.platform()
+
+        if (platform == null) {
+            logger.error("Cannot boot service ${service.name()} because the platform is null. Make sure the group has a valid platform assigned.")
+            return
+        }
+
         service.state = Service.State.STARTING
         service.path.createDirectories()
 
+        val applicationPath = service.group.applicationPlatformFile()
+
+        // download and copy the platform files to the service path
+        platform.prepare(service.group.data.platform.version)
+        Files.copy(applicationPath, service.path.resolve(applicationPath.name), StandardCopyOption.REPLACE_EXISTING)
+
         // basically current only the java command is supported yet
-        val commands = listOf("java", "-jar", "paper.jar")
+        val commands = listOf("java", "-jar", applicationPath.name)
 
         service.process = ProcessBuilder().directory(service.path.toFile()).command(commands).start()
     }
@@ -45,7 +61,7 @@ class LocalRuntimeFactory : RuntimeFactory {
     override fun shutdownApplication(service: Service) {
         //todo shutdown with command in input stream
 
-        if(service.state == Service.State.STOPPING || service.state == Service.State.STOPPING) {
+        if (service.state == Service.State.STOPPING || service.state == Service.State.STOPPING) {
             logger.info("Cannot shutdown service ${service.name()} because it is already stopping or stopped&8. &7Wait for action&8...")
             return
         }
@@ -53,8 +69,10 @@ class LocalRuntimeFactory : RuntimeFactory {
         logger.info("The service &3${service.name()}&7 is now stopping&8...")
         //  if (service.process().waitFor(PolocloudSuite.instance().config().local().processTerminationIdleSeconds(), TimeUnit.SECONDS)) {
 
-        service.process!!.toHandle().destroyForcibly()
-        service.process = null
+        if (service.process != null) {
+            service.process!!.toHandle().destroyForcibly()
+            service.process = null
+        }
 
         Thread.sleep(200) // wait for a process to be destroyed
         service.path.deleteRecursively()
